+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Werewolves are supported! Along with the tags, character sheet, and templates fo
 * Turn off sorting for character listings altogether using `--no_sort`
 * Force listing to contain only body content using `--partial`
 * New setting for default character type in GUI
+* Settings for allowed werewolf auspices and tribes
 
 ### Changed